renmase NoopApiTracer to BaseApiTracer
