+++ gax/src/main/java/com/google/api/gax/tracing/BaseApiTracer.java
@@ -30,25 +30,109 @@
 package com.google.api.gax.tracing;
 
 import com.google.api.core.InternalApi;
+import org.threeten.bp.Duration;
 
 /**
- * Factory that will build {@link ApiTracer}s that do nothing.
+ * A base implementation of {@link ApiTracer} that does nothing.
  *
  * <p>For internal use only.
  */
 @InternalApi("For internal use by google-cloud-java clients only")
-public class NoopApiTracerFactory implements ApiTracerFactory {
-  private static final NoopApiTracerFactory INSTANCE = new NoopApiTracerFactory();
+public class BaseApiTracer implements ApiTracer {
+  private static final ApiTracer INSTANCE = new BaseApiTracer();
 
-  public static NoopApiTracerFactory getInstance() {
+  private static final Scope NOOP_SCOPE =
+      new Scope() {
+        @Override
+        public void close() {
+          // noop
+        }
+      };
+
+  protected BaseApiTracer() {}
+
+  public static ApiTracer getInstance() {
     return INSTANCE;
   }
 
-  protected NoopApiTracerFactory() {}
-
-  /** {@inheritDoc} */
   @Override
-  public ApiTracer newTracer(ApiTracer parent, SpanName spanName, OperationType operationType) {
-    return NoopApiTracer.getInstance();
+  public Scope inScope() {
+    return NOOP_SCOPE;
+  }
+
+  @Override
+  public void operationSucceeded() {
+    // noop
+  }
+
+  @Override
+  public void operationCancelled() {
+    // noop
+  }
+
+  @Override
+  public void operationFailed(Throwable error) {
+    // noop
+  }
+
+  @Override
+  public void connectionSelected(String id) {
+    // noop
+  }
+
+  @Override
+  public void attemptStarted(int attemptNumber) {
+    // noop
+  }
+
+  @Override
+  public void attemptSucceeded() {
+    // noop
+  }
+
+  @Override
+  public void attemptCancelled() {
+    // noop
+  }
+
+  @Override
+  public void attemptFailed(Throwable error, Duration delay) {
+    // noop
+  }
+
+  @Override
+  public void attemptFailedRetriesExhausted(Throwable error) {
+    // noop
+  }
+
+  @Override
+  public void attemptPermanentFailure(Throwable error) {
+    // noop
+
+  }
+
+  @Override
+  public void lroStartFailed(Throwable error) {
+    // noop
+  }
+
+  @Override
+  public void lroStartSucceeded() {
+    // noop
+  }
+
+  @Override
+  public void responseReceived() {
+    // noop
+  }
+
+  @Override
+  public void requestSent() {
+    // noop
+  }
+
+  @Override
+  public void batchRequestSent(long elementCount, long requestSize) {
+    // noop
   }
 }
