+++ gax/src/test/java/com/google/api/gax/tracing/TracedUnaryCallableTest.java
@@ -70,7 +70,7 @@ public class TracedUnaryCallableTest {
 
   @Before
   public void setUp() {
-    parentTracer = NoopApiTracer.getInstance();
+    parentTracer = BaseApiTracer.getInstance();
 
     // Wire the mock tracer factory
     when(tracerFactory.newTracer(
